[Pointer Events] The pointerup, pointerout and pointerleave events may be fired twice https://bugs.webkit.org/show_bug.cgi?id=198028 <rdar://problem/50769425>
Reviewed by Dean Jackson.

Add a new test that checks that we're firing a "pointermove" event when the touch pressure
changes, even when the touch is stationary, and that a single "pointerup" event is fired
as the touch ends and the pressure changes. The relevant code change is done in WebKitAdditions.

* pointerevents/ios/pressure-change-expected.txt: Added.
* pointerevents/ios/pressure-change.html: Added.
* pointerevents/utils.js:
(prototype._handlePointerEvent):
(prototype.stationary):
(prototype._action):

Identifier: 212126@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245506 268f45cc-cd09-0410-ab3c-d52691b4dbfc