[iOS] Layout viewport size on google.com increases after rotating to landscape and back https://bugs.webkit.org/show_bug.cgi?id=198062 <rdar://problem/50547895>
Reviewed by Maciej Stachowiak.

Source/WebKit:

During an animated resize (e.g. when rotating the device on iOS), we currently immediately trigger the new
shrink-to-fit content size heuristic in the middle of dynamicViewportSizeUpdate, after the new view layout size
has been applied to the viewport configuration but before we've issued a resize event to the page.

Thus, on pages that use listen to the resize event and adjust their content accordingly to fit within the new
layout width, we prematurely declare that the page has horizontally overflowed, and try to lay out at a larger
width and scale down. This causes the page to unnecessarily shrink after rotating to landscale orientation and
back.

To fix this, we simply move the call to shrink-to-fit-content to the end of the dynamic viewport size update,
such that the page has had a chance to adjust to the new layout size.

Test: fast/events/ios/rotation/do-not-shrink-to-fit-content-after-rotation.html

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::dynamicViewportSizeUpdate):

LayoutTests:

Add a UIHelper method to simulate device rotation to a given orientation, and use it in a new layout test that
simulates rotation to and from landscape orientation, and verifies that the initial scale did not change from
its expected value of 1.

* fast/events/ios/rotation/do-not-shrink-to-fit-content-after-rotation-expected.txt: Added.
* fast/events/ios/rotation/do-not-shrink-to-fit-content-after-rotation.html: Added.
* resources/ui-helper.js:
(window.UIHelper.rotateDevice.return.new.Promise.):
(window.UIHelper.rotateDevice):
(window.UIHelper):


Identifier: 212154@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245561 268f45cc-cd09-0410-ab3c-d52691b4dbfc