[Pointer Events] Compatibility mouse events can only be prevented while the pointer is pressed https://bugs.webkit.org/show_bug.cgi?id=198178
Patch by Antoine Quint <graouts@apple.com> on 2019-05-23
Reviewed by Dean Jackson.

Source/WebCore:

Test: pointerevents/mouse/compatibility-mouse-events-prevention-mouse-released.html

The Pointer Events spec, in https://www.w3.org/TR/pointerevents/#compatibility-mapping-with-mouse-events, says that "Mouse events
can only be prevented when the pointer is down. Hovering pointers (e.g. a mouse with no buttons pressed) cannot have their mouse
events prevented." We now track whether the pointer is pressed and clear the preventsCompatibilityMouseEvents when the pointer is
moved and it is not pressed.

* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::pointerEventWasDispatched):
* page/PointerCaptureController.h:

LayoutTests:

Add a test that listens "mousemove" events and checks it is not dispatched after preventDefault() is called when handling
 "pointerdown" but it is dispatched after releasing the mouse pointer and moving it again.

* pointerevents/mouse/compatibility-mouse-events-prevention-mouse-released-expected.txt: Added.
* pointerevents/mouse/compatibility-mouse-events-prevention-mouse-released.html: Added.

Identifier: 212248@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245699 268f45cc-cd09-0410-ab3c-d52691b4dbfc