JITOperations getByVal should mark negative array indices as out-of-bounds https://bugs.webkit.org/show_bug.cgi?id=198229
Reviewed by Saam Barati.

JSTests:

* microbenchmarks/get-by-val-negative-array-index.js: Added.
(foo):

Source/JavaScriptCore:

get_by_val with an array or string as base value and a negative index causes DFG to OSR exit,
but baseline doesn't mark it as out-of-bounds, since it only considers positive indices. This
leads to discarding DFG code, recompiling it and exiting at the same bytecode.

This is observed in the prepack-wtb subtest of JetStream2. In popContext#CdOhFJ, the last item
of the array popped and the new last value is accessed using `array[array.length - 1]`, which
is -1 when the array is empty. It shows a ~0.5% progression in JetStream2, but it's within the
noise.

* jit/JITOperations.cpp:
(JSC::getByVal):


Identifier: 212293@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245769 268f45cc-cd09-0410-ab3c-d52691b4dbfc