REGRESSION (r244182) [Mac WK2] Layout Test imported/w3c/web-platform-tests/visual-viewport/viewport-resize-event-on-load-overflowing-page.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=197286
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2019-05-29
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

This is a time sensitive test. It expects to receive a resize event within
two frames after the page load. Scheduling the page update through the
RenderingUpdate made updating the page for the first time happens not
immediately after the page load but after an additional frame.

So we need to wait for this additional frame before checking whether the
'resize' event was fired.

* web-platform-tests/visual-viewport/viewport-resize-event-on-load-overflowing-page.html:

LayoutTests:

* platform/mac-wk2/TestExpectations:

Identifier: 212385@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245877 268f45cc-cd09-0410-ab3c-d52691b4dbfc