Rare crash under `WebPage::shouldIgnoreMetaViewport const` when shrinking to fit content https://bugs.webkit.org/show_bug.cgi?id=198374 <rdar://problem/51261348>
Reviewed by Tim Horton.

This is a speculative fix for the crash; what appears to be happening is that the shrink-to-fit-content timer
is somehow being scheduled after the page has already been closed. We can probably fix this by bailing when
scheduling shrink-to-fit-content if m_isClosed is true.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::scheduleShrinkToFitContent):
(WebKit::WebPage::immediatelyShrinkToFitContent):


Identifier: 212401@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245894 268f45cc-cd09-0410-ab3c-d52691b4dbfc