Fix 32-bit/64-bit mismatch in PointerCaptureController::elementWasRemoved https://bugs.webkit.org/show_bug.cgi?id=198501 <rdar://problem/51370464>
Reviewed by Chris Dumez.

keyAndValue.key is assigned to pointerId. KeyAndValue.key is a
int64_t, whereas pointerId is a PointerID, aka int32_t. This mismatch
is normally just a warning, but breaks builds where warnings are
treated as errors.

This issue is not encountered in most builds because the warning is
disabled in the majority of build configurations. But there are some
where the warning is not disabled, and so those builds break.

Address this conversion error/warning by explicitly casting
keyAndValue.key to a PointerID (and adding a debug check to make sure
the cast is OK).

No new tests -- no new functionality.

* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::elementWasRemoved):


Identifier: 212533@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246072 268f45cc-cd09-0410-ab3c-d52691b4dbfc